+++ packages/snaps-rpc-methods/src/restricted/getBip32Entropy.test.ts
@@ -75,11 +75,11 @@ describe('getBip32EntropyImplementation', () => {
     });
 
     messenger.registerActionHandler(
-      'KeyringController:withKeyring',
+      'KeyringController:withKeyringV2Unsafe',
       async (_selector, operation) =>
         operation({
           keyring: {
-            type: 'HD Key Tree',
+            type: 'hd',
             mnemonic: TEST_SECRET_RECOVERY_PHRASE_BYTES,
             seed: TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
           },
@@ -247,7 +247,7 @@ describe('getBip32EntropyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyring',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );
@@ -287,7 +287,7 @@ describe('getBip32EntropyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyring',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );
@@ -341,11 +341,11 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyring',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );
@@ -371,11 +371,11 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyring',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );
@@ -401,11 +401,11 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyring',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );
@@ -435,11 +435,11 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyring',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
-              type: 'HD Key Tree',
+              type: 'hd',
             },
           }),
       );