+++ packages/snaps-rpc-methods/src/restricted/getEntropy.test.ts
@@ -20,7 +20,7 @@ describe('getEntropyBuilder', () => {
         getUnlockPromise: true,
         getClientCryptography: true,
       },
-      actionNames: ['KeyringController:withKeyring'],
+      actionNames: ['KeyringController:withKeyringV2Unsafe'],
     });
   });
 
@@ -55,11 +55,11 @@ describe('getEntropyImplementation', () => {
     });
 
     messenger.registerActionHandler(
-      'KeyringController:withKeyring',
+      'KeyringController:withKeyringV2Unsafe',
       async (_selector, operation) =>
         operation({
           keyring: {
-            type: 'HD Key Tree',
+            type: 'hd',
             seed: TEST_SECRET_RECOVERY_PHRASE_SEED_BYTES,
           },
         }),
@@ -133,7 +133,7 @@ describe('getEntropyImplementation', () => {
     );
 
     expect(messenger.call).toHaveBeenCalledWith(
-      'KeyringController:withKeyring',
+      'KeyringController:withKeyringV2Unsafe',
       { id: 'source-id' },
       expect.any(Function),
     );